requirements.txt
python-gitlab
coverage
twine